backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
mysqlclient==2.2.0
python-dotenv==1.0.0
python-decouple==3.8
redis==5.0.1
cassandra-driver==3.28.0
gunicorn==21.2.0
Pillow==10.1.0
boto3==1.28.65
django-storages==1.14.2
requests==2.31.0
celery==5.3.4
pyyaml==6.0.1
django-filter==23.3
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
